Better use of the WriterT combiner.
Figured out a better way to write code that works with the combined
Writer/Maybe monad, by using more generic findFirst which works on any
Alternative type, rather than findJust which was specific to Maybe. This
yields code that I think is closer to the intent of using combined
monads.  I'm still not clear that it's worth the trouble in this case
though.